Changed the service to one chatroom
Ran into various problems* keeping track of users in multiple chatrooms
across multiple devices/tabs. Switching to a single chatroom in order
to allow for the most robust feature set possible.

*[mizzao:user-status has comprehensive server-side functionality, but
does not allow for the access of the connection._id client-side (which
would be required for removing the user from a specific chatroom upon
unload/close).]